Refactor: QuestionModel 위치변경
