+++ src/main/java/supernova/whokie/question/controller/QuestionController.java
@@ -15,7 +15,7 @@ import supernova.whokie.global.annotation.Authenticate;
 import supernova.whokie.global.dto.GlobalResponse;
 import supernova.whokie.global.dto.PagingResponse;
 import supernova.whokie.group_member.controller.dto.GroupMemberResponse;
-import supernova.whokie.question.controller.dto.QuestionModel;
+import supernova.whokie.question.service.dto.QuestionModel;
 import supernova.whokie.question.controller.dto.QuestionRequest;
 import supernova.whokie.question.controller.dto.QuestionResponse;
 import supernova.whokie.question.service.QuestionService;
