+++ src/main/java/supernova/whokie/question/service/QuestionService.java
@@ -9,7 +9,7 @@ import supernova.whokie.friend.Friend;
 import supernova.whokie.friend.infrastructure.repository.FriendRepository;
 import supernova.whokie.global.exception.EntityNotFoundException;
 import supernova.whokie.question.Question;
-import supernova.whokie.question.controller.dto.QuestionModel;
+import supernova.whokie.question.service.dto.QuestionModel;
 import supernova.whokie.question.repository.QuestionRepository;
 import supernova.whokie.user.Users;
 import supernova.whokie.user.controller.dto.UserResponse;
